Add Farm. Rename modal files.
